#271 Add vertical spacing between the two related-action buttons on narrow screens.
On narrow screens instead of being side by side they're stacked. The distance between the 'Add to clipboard' button and the related actions div is 0.75rem, so I think the only stacked distance between the two related actions buttons that looks good is 0.75rem.  With this fix that changes the distance from the bottom of the related actions button(s) to the next field to be 0.75rem instead of the 0.5rem that it was previously.