#123 Display a 'Zap spammer' button on profile pages for spam zappers.
I think this will be a little different than current BugGuide, where (I'm guessing) the spam zapping text is added to the end of the '# comments contributed by this user' text, which exists at the end of the license section - here that comments text (which currently doesn't exist on beta, but may be added later) is/will be displayed separately from the license info. Since the license info is the only field guaranteed to have content, we stick with adding the spam info to the license field, and just repeat the comment info there.