#123 Implement the 'zap/%' path.
The previous version of this code didn't unset the spammer's user status; here that was required in order for the spammer's profile page to no longer be visible (except to admin).

An even bigger change was to switch to updating nodes using node_load/node_save instead of updating the database directly. There were a couple issues that prompted that:
* if you only update status in the node table, you see no evident change on the website; you must also update status on the node_revision table, even though we're not using revisions.
* the trickier issue though, is that if you compare what "Cancel the selected user" in the admin panel and then "Disable the account and unpublish its content" does to what we were doing by hand, in addition to node and node_revision being updated, tracker_node and tracker_user will be updated as well. I couldn't see any difference when we didn't update those tables when doing things by hand, but eventually we may have revisions on spammy articles, and then it could become an issue. And updating those tables by hand seems a little tricky - sometimes you remove a row entirely, sometimes you just change its 'published' column (seemingly depending on whether the spammer created the node or just edited it somehow, but I'm not sure about that).
The good news is that we can still undo the changes (and that doing this through node seems to be the canonical way to do it), the bad news is that doing so requires running a drush script instead of updating the databse directly (or maybe that's good news too and we should put all of the undo steps in a drush script).

The spammer's profile page isn't visible (except to admin) after being zapped, but user/%spammer_id/images, user/%spammer_id/content, and user/%spammer_id/comments still show the unpublished images, nodes, and (some of?) the comments (maybe just the comments that don't appear on unpublished images?). If we want to disable those as well I think we can do so in a followup.

Once a user is zapped, spamzapper and authenticated (and editors etc) can no longer see the user's profile, comments, images, links, or refs. Admin sees images, links, and refs in pink (unpublished), but spam comments appear no differently (I ran across a fixed core bug about this, I wonder if we have the fix). The profile page of the spammer appears unchanged for Admin, except that in the spamzapper section of that profile, all content is reported as unpublished (so admin can recognize zapped users even on their profile page).

A zapped user's username still autocompletes in the user field of advanced search, but searches return no results; a basic search doesn't yeild any results as far as I can tell.