#304 Change the type of field_user_homepage to link.
It had been text. This also changes the default for that field from 'http:' to ''.

The behavior for display here is slightly different than in current bugguide. In current bugguide if you enter a url without a protocol, the url is displayed with no protocol, as entered by the user. With this commit, if the user enters a url without a protocol, Drupal displays it with a protocol of 'http://'. If you enter a url with a protocol of 'https://' then Drupal keeps that protocol. I presume this is kind of a "hey, remember that the protocol by default is http://, and not https:// unless you specify that" kind of a thing, which I guess is good. There is a display option of 'URL with the protocol removed', but then the protocol is _always_ removed, even if the user specified one when they entered the url.

(For the record, the type was changed by deleting the old field, creating a new one with the old field name but a type of 'link', and then exporting the bguserfields feature and diffing to figure out what to include in this commit. I'm not 100% certain, but I don't think changing the field type in this case should have any impact on migration.)