#281 Honor an empty date when creating a cloned image.
Without this patch if you clone an image with an empty date, the current date is inserted in the date field of the cloned node create node form.

The cloned node in this case correctly has an empty field_bgimage_date, but it seems the form creation code in the clone case inserts a default date in that case (I'm not sure why it only happens on cloned nodes and not when editing an existing node with no date field set; maybe just using a different form generation path in the two cases?). The #default_value being set here is the same as that seen on the edit form of an existing node with no date set.