#367 Add a 'Recent images' action for displaying recent images for a taxon
'options' on a MENU_LOCAL_ACTION are ignored (according to the docs), and my initial attempts at adding a CSS class for the action using various hooks failed, so we're just doing it in CSS directly.

My (vague) understanding is that in order for a local action to display on a given page, the path of the action must extend the path of the page (or something along those lines); so here our recent taxon images action path, 'node/%bgpage_node/bgimage/recent', extends the images tab path 'node/%bgpage_node/bgimage'. (This is a different path than what current BugGuide uses for recent taxon images.)

The title we were setting in our new action was being overwritten by bg_preprocess_page; as it turns out, if we want our taxon name to be markuped then the place it was being overwritten is the only place it can be set anyway, and the title being set there is the non-constant part of the title we wanted anyway, so I guess it could have been worse.