Revert "#281 Honor an empty date when creating a cloned image."
This reverts commit 196dd9c193cea394d3ab595d33ff35b1850f4330.
When I created that commit I didn't realize the non-empty date was a result of a default date of 'today' set on the bgimage date field; we now have no default date (as on current BugGuide), so there's no need for that fix. Also, for reasons unkown to me, the commit being reverted never fixed the problem on beta, only locally, so I'm glad for another fix!