#130 Factor out a validation function for bgpage nodes
When a user uses the quickadd form to create a new guide page they're redirected to the new guide page to finish filling out the page data. But there's no guarantee that they'll do that, so we want full validation on the fields they do fill out on the quick add form so that we're left in a sane state - for example we wouldn't want them to quickadd a taxon that already exists, or to misspell 'idae', etc.

Note that we're no longer checking isset on $form_state values like $form_state['values']['field_scientific_name'][LANGUAGE_NONE][0]['value'] since it seems they're always set (though they may be empty).

At the end of bgpage_node_form_validate we remove the variables being passed on to the submit function since only node is processing the submit and it's not going to use those variables.