#130 Port the BugGuide 6 bgpage quickadd code
The code has been copied from

https://bitbucket.org/isubit/bugguide6/src/88f82e35f248eac41de7aa36c61ac5558bc7a1c4/sites/all/modules/bgpage/bgpage.module#lines-810:904

and

https://bitbucket.org/isubit/bugguide6/src/88f82e35f248eac41de7aa36c61ac5558bc7a1c4/sites/all/modules/bgpage/bgpage.module#lines-1163:1204

and then updated for Drupal 7.

In the quickadd submit function we state that the validate function always sets a common name - that's not true yet, but it will be after a future commit in this series.

The validation function was updated to match the criteria listed at #130. Those validations will be merged with the existing bgpage_node_form_validate in a followup commit.