#159 Close the bgpage-taxon-desc span in the taxonomy layout
We were failing to close the spans with class bgpage-taxon-desc, and as a result we were nesting the taxonomy layout when we didn't intend to.

This is a somewhat speculative fix (I'm not on beta and I haven't created a taxonomy layout locally that's long enough to trigger the bug), but in firefox if I look at the first incorrectly layed out line in view source the link shows up in red and if I hover my mouse over the red text it says 'The document tree is too deep. The tree will be flattened to 513 elements deep.'. [I thought current BugGuide actually used nested tags for its taxonomy layout, so either I'm incorrect about that, or it nests in a different way, or that particular browser error only gets triggered when there are unclosed nested tags.]