#21 Add validation and submit handling for guide page sequencing
This is essentially ported from BG6; the only addition is that we now validate that all sequence values are unique - that will mean that any migrated sequences with repeated values will be forced to be updated the next time they're edited.

We're not using a transaction here; if weights are only partially updated that's not a big deal and easy to fix (... right?).